ReplenishBillService修改

+++ src/com/casesoft/dmc/model/logistics/vo/ReplenishSkuVO.java
@@ -1,5 +1,6 @@
 package com.casesoft.dmc.model.logistics.vo;
 
+import com.casesoft.dmc.core.util.reflection.ConvertUtils;
 import io.swagger.models.auth.In;
 
 import java.io.Serializable;
@@ -19,13 +20,13 @@ public class ReplenishSkuVO implements Serializable {
     private String sizeId;
     private String sizeName;
     private List<ReplenishCodeVO> codeVOList;
-    private Integer skuTotQty;
+    private Long skuTotQty;
     private Integer skuTotActConvertQty;
-    private Integer skuTotInstockQty;
+    private Long skuTotInstockQty;
 
     public ReplenishSkuVO(){}
 
-    public ReplenishSkuVO(String sku, String styleId, String colorId, String sizeId, Integer skuTotQty, Integer skuTotActConvertQty, Integer skuTotInstockQty){
+    public ReplenishSkuVO(String sku, String styleId, String colorId, String sizeId, Long skuTotQty, Integer skuTotActConvertQty, Long skuTotInstockQty) {
         this.sku = sku;
         this.styleId = styleId;
         this.colorId = colorId;
@@ -99,27 +100,16 @@ public class ReplenishSkuVO implements Serializable {
         this.codeVOList = codeVOList;
     }
 
-    public Integer getSkuTotQty() {
-        return skuTotQty;
-    }
+    public Long getSkuTotQty() { return skuTotQty; }
 
-    public void setSkuTotQty(Integer skuTotQty) {
-        this.skuTotQty = skuTotQty;
-    }
+    public void setSkuTotQty(Long skuTotQty) { this.skuTotQty = skuTotQty; }
 
-    public Integer getSkuTotActConvertQty() {
-        return skuTotActConvertQty;
-    }
+    public Integer getSkuTotActConvertQty() { return skuTotActConvertQty; }
 
-    public void setSkuTotActConvertQty(Integer skuTotActConvertQty) {
-        this.skuTotActConvertQty = skuTotActConvertQty;
-    }
+    public void setSkuTotActConvertQty(Integer skuTotActConvertQty) { this.skuTotActConvertQty = skuTotActConvertQty; }
 
-    public Integer getSkuTotInstockQty() {
-        return skuTotInstockQty;
-    }
+    public Long getSkuTotInstockQty() { return skuTotInstockQty; }
 
-    public void setSkuTotInstockQty(Integer skuTotInstockQty) {
-        this.skuTotInstockQty = skuTotInstockQty;
-    }
+    public void setSkuTotInstockQty(Long skuTotInstockQty) { this.skuTotInstockQty = skuTotInstockQty; }
+
 }
